+++ Practica_IA_19_20/src/practica/objetos/Informacion.java
@@ -14,10 +14,10 @@ public abstract class Informacion {
             {25, 20, 15, 15, 10, 10, 5, 5, 0},
     };
 
-    public static byte getCoste(Areas Origen, Areas Destino, int peso) {
-        byte fila = getIndex(Origen) , columna = getIndex(Destino);
+    public static double getCoste(Areas Origen, Areas Destino, double peso) {
+        int fila = getIndex(Origen) , columna = getIndex(Destino);
         if(!(fila == -1 || columna == -1))
-            return (byte)(costes[fila][columna]*((5+peso)/5));
+            return (double)(costes[fila][columna]*((5+peso)/5));
         return -1;
     }
     private static byte getIndex(Areas Area){
